Refactored and pulled annotations into separate compostable
